+++ src/Faker/Game/Touhou.hs
@@ -1,0 +1,19 @@
+{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE OverloadedStrings #-}
+
+module Faker.Game.Touhou where
+
+import Data.Text (Text)
+import Faker (Fake(..))
+import Faker.Provider.Touhou
+import Faker.TH
+
+$(generateFakeField "touhou" "games")
+
+$(generateFakeField "touhou" "characters")
+
+$(generateFakeField "touhou" "spell_cards")
+
+$(generateFakeField "touhou" "locations")
+
+$(generateFakeField "touhou" "songs")